requirements.txt
Django==5.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
python-dotenv==1.0.0
Pillow==10.1.0
reportlab==4.0.7  # For PDF export
openai==1.3.7  # For AI recipe suggestions
WeasyPrint==60.1  # For PDF generation
django-cors-headers==4.3.1  # For frontend integration
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
accelerate==0.25.0
